refactor lfc and lff with a CliBase class
